+++ src/functions/event-topic/index.ts
@@ -1,5 +1,6 @@
 export { EventTopicBackfillForAll } from './backfill.js';
 export { EventTopicCleanBackfillForAll } from './clean-backfill.js';
+export { EventTopicGenerateCodeForJsonSchema } from './generate-code-json-schema.js';
 export { EventTopicGenerateCodeReferencedInProjectForAll } from './generate-code-referenced-in-project.js';
 export { EventTopicListReferencedInProjectForServerlessFunctions } from './list-referenced-in-project-serverless-functions.js';
 export { EventTopicListReferencedInProjectForServiceContainer } from './list-referenced-in-project-service-container.js';